+++ server/src/main/java/org/opensearch/index/search/stats/SearchStats.java
@@ -35,6 +35,7 @@ package org.opensearch.index.search.stats;
 import org.opensearch.Version;
 import org.opensearch.common.Nullable;
 import org.opensearch.common.unit.TimeValue;
+import org.opensearch.common.util.FeatureFlags;
 import org.opensearch.core.common.Strings;
 import org.opensearch.core.common.io.stream.StreamInput;
 import org.opensearch.core.common.io.stream.StreamOutput;
@@ -70,6 +71,7 @@ public class SearchStats implements Writeable, ToXContentFragment {
         private long concurrentQueryCount;
         private long concurrentQueryTimeInMillis;
         private long concurrentQueryCurrent;
+        private long queryConcurrency;
 
         private long fetchCount;
         private long fetchTimeInMillis;
@@ -98,6 +100,7 @@ public class SearchStats implements Writeable, ToXContentFragment {
             long concurrentQueryCount,
             long concurrentQueryTimeInMillis,
             long concurrentQueryCurrent,
+            long queryConcurrency,
             long fetchCount,
             long fetchTimeInMillis,
             long fetchCurrent,
@@ -118,6 +121,7 @@ public class SearchStats implements Writeable, ToXContentFragment {
             this.concurrentQueryCount = concurrentQueryCount;
             this.concurrentQueryTimeInMillis = concurrentQueryTimeInMillis;
             this.concurrentQueryCurrent = concurrentQueryCurrent;
+            this.queryConcurrency = queryConcurrency;
 
             this.fetchCount = fetchCount;
             this.fetchTimeInMillis = fetchTimeInMillis;
@@ -163,6 +167,7 @@ public class SearchStats implements Writeable, ToXContentFragment {
                 concurrentQueryCount = in.readVLong();
                 concurrentQueryTimeInMillis = in.readVLong();
                 concurrentQueryCurrent = in.readVLong();
+                queryConcurrency = in.readVLong();
             }
         }
 
@@ -174,6 +179,7 @@ public class SearchStats implements Writeable, ToXContentFragment {
             concurrentQueryCount += stats.concurrentQueryCount;
             concurrentQueryTimeInMillis += stats.concurrentQueryTimeInMillis;
             concurrentQueryCurrent += stats.concurrentQueryCurrent;
+            queryConcurrency += stats.queryConcurrency;
 
             fetchCount += stats.fetchCount;
             fetchTimeInMillis += stats.fetchTimeInMillis;
@@ -213,6 +219,7 @@ public class SearchStats implements Writeable, ToXContentFragment {
             pitCount += stats.pitCount;
             pitTimeInMillis += stats.pitTimeInMillis;
             pitCurrent += stats.pitCurrent;
+            queryConcurrency += stats.queryConcurrency;
         }
 
         public long getQueryCount() {
@@ -237,6 +244,14 @@ public class SearchStats implements Writeable, ToXContentFragment {
 
         public TimeValue getConcurrentQueryTime() {
             return new TimeValue(concurrentQueryTimeInMillis);
+        }
+
+        public double getConcurrentAvgSliceCount() {
+            if (concurrentQueryCount == 0) {
+                return 0;
+            } else {
+                return queryConcurrency / (double) concurrentQueryCount;
+            }
         }
 
         public long getConcurrentQueryTimeInMillis() {
@@ -343,6 +358,7 @@ public class SearchStats implements Writeable, ToXContentFragment {
                 out.writeVLong(concurrentQueryCount);
                 out.writeVLong(concurrentQueryTimeInMillis);
                 out.writeVLong(concurrentQueryCurrent);
+                out.writeVLong(queryConcurrency);
             }
         }
 
@@ -352,9 +368,12 @@ public class SearchStats implements Writeable, ToXContentFragment {
             builder.humanReadableField(Fields.QUERY_TIME_IN_MILLIS, Fields.QUERY_TIME, getQueryTime());
             builder.field(Fields.QUERY_CURRENT, queryCurrent);
 
-            builder.field(Fields.CONCURRENT_QUERY_TOTAL, concurrentQueryCount);
-            builder.humanReadableField(Fields.CONCURRENT_QUERY_TIME_IN_MILLIS, Fields.CONCURRENT_QUERY_TIME, getConcurrentQueryTime());
-            builder.field(Fields.CONCURRENT_QUERY_CURRENT, concurrentQueryCurrent);
+            if (FeatureFlags.isEnabled(FeatureFlags.CONCURRENT_SEGMENT_SEARCH)) {
+                builder.field(Fields.CONCURRENT_QUERY_TOTAL, concurrentQueryCount);
+                builder.humanReadableField(Fields.CONCURRENT_QUERY_TIME_IN_MILLIS, Fields.CONCURRENT_QUERY_TIME, getConcurrentQueryTime());
+                builder.field(Fields.CONCURRENT_QUERY_CURRENT, concurrentQueryCurrent);
+                builder.field(Fields.CONCURRENT_AVG_SLICE_COUNT, getConcurrentAvgSliceCount());
+            }
 
             builder.field(Fields.FETCH_TOTAL, fetchCount);
             builder.humanReadableField(Fields.FETCH_TIME_IN_MILLIS, Fields.FETCH_TIME, getFetchTime());
@@ -484,6 +503,7 @@ public class SearchStats implements Writeable, ToXContentFragment {
         static final String CONCURRENT_QUERY_TIME = "concurrent_query_time";
         static final String CONCURRENT_QUERY_TIME_IN_MILLIS = "concurrent_query_time_in_millis";
         static final String CONCURRENT_QUERY_CURRENT = "concurrent_query_current";
+        static final String CONCURRENT_AVG_SLICE_COUNT = "concurrent_avg_slice_count";
         static final String FETCH_TOTAL = "fetch_total";
         static final String FETCH_TIME = "fetch_time";
         static final String FETCH_TIME_IN_MILLIS = "fetch_time_in_millis";